@@ -17,10 +17,10 @@ class ChallengesController < ApplicationController
     if @challenge.save
       Rails.logger.info "CHALLENGE_LOG: #{current_user.full_name} successfully challenged the #{spot&.to_s} spot"
       send_challenge_success_email
-      render :show, status: 201
+      render :show, status: :created
     else
       Rails.logger.info "CHALLENGE_LOG: #{current_user.full_name} couldn't challenge #{spot&.to_s}. Errors: #{@challenge.errors}"
-      render json: { resource: 'challenge', errors: @challenge.errors }, status: 409
+      render json: { resource: 'challenge', errors: @challenge.errors }, status: :conflict
     end
   end
   # rubocop:enable Metrics/MethodLength, Metrics/LineLength
@@ -66,13 +66,13 @@ class ChallengesController < ApplicationController
 
   def ensure_not_challenging_alternate!
     return if params[:spot][:file] < 13
-    render json: { resource: 'challenge', errors: [challenge: 'can\'t challenge alternate'] }, status: 403
+    render json: { resource: 'challenge', errors: [challenge: 'can\'t challenge alternate'] }, status: :forbidden
   end
 
   def ensure_performance_is_current_and_open!
     next_performance = Performance.next
     return if next_performance&.window_open?
-    render json: { resource: 'challenge', errors: [performance: 'window not open'] }, status: 403
+    render json: { resource: 'challenge', errors: [performance: 'window not open'] }, status: :forbidden
   end
 
   def ensure_user_has_not_already_made_challenge!
@@ -80,7 +80,7 @@ class ChallengesController < ApplicationController
     p = Performance.next
     challenges = user.challenges
     return unless challenges.any? { |c| c.performance.id == p.id }
-    render json: { resouce: 'challenge', errors: [user: 'can\'t make more than one challenge'] }, status: 403
+    render json: { resouce: 'challenge', errors: [user: 'can\'t make more than one challenge'] }, status: :forbidden
   end
 
   def ensure_user_is_challenging_correct_instrument_and_part!
@@ -92,7 +92,7 @@ class ChallengesController < ApplicationController
         resource: 'challenge',
         errors: [user: 'can\'t challenge someone of a different instrument or part']
       },
-      status: 403
+      status: :forbidden
     )
   end
 
@@ -102,7 +102,7 @@ class ChallengesController < ApplicationController
     challenge = Challenge.find_by(spot: spot, performance: performance)
     return if challenge.nil?
     return if challenge.open_spot_challenge_type? && challenge.users.length < 2
-    render json: { resource: 'challenge', errors: [spot: 'spot has already been challenged'] }, status: 403
+    render json: { resource: 'challenge', errors: [spot: 'spot has already been challenged'] }, status: :forbidden
   end
 
   def ensure_user_can_submit_for_approval!
