@@ -7,9 +7,9 @@ class DisciplineActionsController < ApplicationController
   def create
     @da = DisciplineAction.new create_params
     if @da.save
-      render :show, status: 201
+      render :show, status: :created
     else
-      render json: { resource: 'discipline_action', errors: @da.errors }, status: 409
+      render json: { resource: 'discipline_action', errors: @da.errors }, status: :conflict
     end
   end
 
@@ -33,7 +33,7 @@ class DisciplineActionsController < ApplicationController
     p = Performance.find_by id: params[:discipline_action][:performance_id]
     p = DisciplineAction.find_by(id: params[:id])&.performance if p.nil?
     return if !p.nil? && !p.stale?
-    render json: { resource: 'discipline_action', errors: [{ performance: 'can\'t be expired' }] }, status: 403
+    render json: { resource: 'discipline_action', errors: [{ performance: 'can\'t be expired' }] }, status: :forbidden
   end
 
   # rubocop:disable Metrics/LineLength
@@ -47,7 +47,7 @@ class DisciplineActionsController < ApplicationController
     render json: {
       resource: 'discipline_action',
       errors: [{ discipline_action: "#{user.first_name} #{user.last_name}'s spot has already been challenged. Please remove that challenge before deleting this discipline action." }]
-    }, status: 403
+    }, status: :forbidden
   end
   # rubocop:enable Metrics/LineLength
 end
