@@ -9,10 +9,10 @@ class PasswordResetRequestsController < ApplicationController
     @prr = PasswordResetRequest.new user: @user
     # We're reloading because prr.expires is a default value created in the db. Unfortunately, it doesn't get returned
     if @prr.save && @prr.reload
-      render :create, status: 201
+      render :create, status: :created
       _send_email
     else
-      render json: { resource: 'PasswordResetRequest', errors: @prr.errors }, status: 409
+      render json: { resource: 'PasswordResetRequest', errors: @prr.errors }, status: :conflict
     end
   end
   # rubocop:enable Metrics/MethodLength
